requirements.txt
streamlit
pandas
bcrypt
geopy
folium
streamlit-folium
streamlit-calendar
googlemaps
holidays
streamlit-gsheets-connection